Quintic interpolation moved to separate section
Perlin's "Improved Noise" with quintic interpolation is separate from Simplex Noise, which does not even use interpolation. The contributions from each simplex corner are added, not blended. I created a separate section "Improved Noise" to hold the explanation of the quintic interpolation polynomial (which is not a Hermite polynomial, btw).